Windows build and warning fixes

+++ Engine/Core/src/OpenGL/VertexArray.cpp
@@ -29,7 +29,7 @@ namespace VEX
 			const auto& element = elements[i];
 
 			glEnableVertexAttribArray(i);
-			glVertexAttribPointer(i, element.count, element.type, element.normalized, layout.GetStride(), (const void*) offset);
+			glVertexAttribPointer(i, element.count, element.type, element.normalized, layout.GetStride(), (const void*) (uintptr_t) offset);
 
 			offset += element.count * VertexBufferElement::GetSizeOfType(element.type);
 		}
